Queue

+++ service/iplayerService.js
@@ -9,9 +9,9 @@ import socketService from './socketService.js';
 import loggingService from './loggingService.js';
 import sonarrService from './sonarrService.js';
 import NodeCache from 'node-cache';
+import queueService from './queueService.js';
 
 const downloads = {};
-const processes = {};
 const timestampFile = 'iplayarr_timestamp';
 
 const episodeRegex = /([0-9]+:)[^a-zA-Z]([^,]+),[^a-zA-Z]([^,]+),[^a-zA-Z]([^,]+)(?:$|\n)/;
@@ -72,18 +72,22 @@ const iplayerService = {
     },
 
     cancel: async (id) => {
-        processes[id].kill('SIGINT');
-        delete processes[id];
-        const {uuid} = Object.values(downloads).find((d) => d.id === id)
-        delete downloads[uuid];
+        queueService.cancelItem(id);
 
-        const downloadDir = getParameter("DOWNLOAD_DIR");
-        const uuidPath = path.join(downloadDir, uuid);
 
-        try{
-            fs.rmSync(uuidPath, { recursive: true, force: true });
-        } catch (err) {
-           loggingService.error(`Error deleting ${uuidPath}:`, err);
+        const download = Object.values(downloads).find((d) => d.id === id)
+        if (download){
+            const {uuid} = download;
+            delete downloads[uuid];
+
+            const downloadDir = getParameter("DOWNLOAD_DIR");
+            const uuidPath = path.join(downloadDir, uuid);
+
+            try{
+                fs.rmSync(uuidPath, { recursive: true, force: true });
+            } catch (err) {
+            loggingService.error(`Error deleting ${uuidPath}:`, err);
+            }
         }
     },
 
@@ -102,8 +106,6 @@ const iplayerService = {
         loggingService.debug(`Executing get_iplayer with args: ${allArgs.join(" ")}`);
         const downloadProcess = spawn(exec, allArgs);
 
-        processes[id] = downloadProcess;
-
         const download = {
             uuid,
             id,
@@ -118,33 +120,35 @@ const iplayerService = {
         downloads[uuid] = download;
 
         downloadProcess.stdout.on('data', (data) => {
-            socketService.emit('log', {id, message : data.toString(), timestamp : new Date()});
-            console.log(data.toString());
-            const lines = data.toString().split("\n");
-            const filenameLine = lines.find((l) => filenameRegex.exec(l));
-            if (filenameLine) {
-                const filenameMatch = filenameRegex.exec(filenameLine);
-                let filename = `${filenameMatch[1]} ${filenameMatch[2]}`;
-                const cachedFilename = filenameCache.get(id);
-                filename = (cachedFilename || legacyCreateNZBName(filename)) + ".mp4";
-                downloads[uuid].filename = filename;
-            }
-            const progressLines = lines.filter((l) => progressRegex.exec(l));
-            if (progressLines.length > 0) {
-                const progressLine = progressLines.pop();
-                const match = progressRegex.exec(progressLine);
-                const [_, progress, size, speed, eta] = match;
-                downloads[uuid].progress = parseFloat(progress);
-                downloads[uuid].size = parseFloat(size);
-                downloads[uuid].speed = parseFloat(speed);
-                downloads[uuid].eta = eta;
+            if (downloads[uuid]){
+                socketService.emit('log', {id, message : data.toString(), timestamp : new Date()});
+                console.log(data.toString());
+                const lines = data.toString().split("\n");
+                const filenameLine = lines.find((l) => filenameRegex.exec(l));
+                if (filenameLine) {
+                    const filenameMatch = filenameRegex.exec(filenameLine);
+                    let filename = `${filenameMatch[1]} ${filenameMatch[2]}`;
+                    const cachedFilename = filenameCache.get(id);
+                    filename = (cachedFilename || legacyCreateNZBName(filename)) + ".mp4";
+                    downloads[uuid].filename = filename;
+                }
+                const progressLines = lines.filter((l) => progressRegex.exec(l));
+                if (progressLines.length > 0) {
+                    const progressLine = progressLines.pop();
+                    const match = progressRegex.exec(progressLine);
+                    const [_, progress, size, speed, eta] = match;
+                    downloads[uuid].progress = parseFloat(progress);
+                    downloads[uuid].size = parseFloat(size);
+                    downloads[uuid].speed = parseFloat(speed);
+                    downloads[uuid].eta = eta;
 
-                const percentFactor = (100 - parseFloat(progress)) / 100;
-                const sizeLeft = downloads[uuid].size * percentFactor;
-                downloads[uuid].sizeLeft = sizeLeft;
-            }
+                    const percentFactor = (100 - parseFloat(progress)) / 100;
+                    const sizeLeft = downloads[uuid].size * percentFactor;
+                    downloads[uuid].sizeLeft = sizeLeft;
+                }
 
-            socketService.emit('downloads', Object.values(downloads));
+                queueService.updateQueue(id, downloads[uuid]);
+            }
         });
 
         downloadProcess.on('close', async (code) => {
@@ -170,8 +174,10 @@ const iplayerService = {
                 }
             }
             delete downloads[uuid];
-            delete processes[id];
+            queueService.removeFromQueue(id);
         });
+
+        return downloadProcess;
     },
 
     search: (term, season, episode) => {
